finished auth for order routes

+++ server/app/routes/orders/index.js
@@ -58,16 +58,12 @@ router.get('/:orderid', function (req, res, next){
 		.exec()
 		.then(
 			function (user){
-				if (_.includes(user.orders, req.params.orderid))
-					Order.findById(req.params.orderid).exec()
-					.then(
-						function (order){
-							res.json(order);
-						},
-						function (err){
-							next(err);
-						}
-					);
+				if (_.includes(user.orders, req.params.orderid)) {
+					var theOrder = _.find(user.orders, {_id: req.params.orderid})
+					res.json(theOrder);
+				} else {
+					res.sendStatus(403).end();
+				}
 			},
 			function (err){
 				next(err);
@@ -86,26 +82,51 @@ router.post('/', function (req, res, next){
 
 //edits this order
 router.put('/:orderid', function (req, res, next){
-	Order.findByIdAndUpdate(req.params.orderid, req.body).exec()
-	.then(
-		function (order){
-			res.status(200).send(order);
-		},
-		function (err){
+	if (req.user.admin){
+		Order.findById(req.params.orderid)
+		.exec()
+		.then(function (order) {
+			order.changeStatus(req.body.orderstatus);
+			order.save();
+			res.json(order);
+			console.log(order);
+		}, function(err){
 			next(err);
-		}
-	);
+		});
+
+	} else {
+		User.findById(req.user._id)
+		.populate('orders')
+		.exec()
+		.then(
+			function (user){
+				if (_.includes(user.orders, req.params.orderid)) {
+					var theOrder = _.find(user.orders, {_id: req.params.orderid});
+					theOrder.cancelOrder();
+					theOrder.save();
+					res.json(theOrder);
+				} else {
+					res.sendStatus(403).end();
+				}
+			},
+			function (err){
+				next(err);
+			}
+		);
+	}
 });
 
 // delete this order
 router.delete('/:orderid', function (req, res, next){
-	Order.findByIdAndRemove(req.params.orderid).exec()
-	.then(
-		function (){
-			res.status(204).send();
-		},
-		function (err){
-			next(err);
-		}
-	);
+	if (req.user.admin){
+		Order.findByIdAndRemove(req.params.orderid).exec()
+		.then(
+			function (){
+				res.status(204).send();
+			},
+			function (err){
+				next(err);
+			}
+		);
+	}
 });
